requirements.txt
streamlit
numpy
keras
tensorflow-cpu
h5py==2.10.0
pillow
